fix(test) + feat(platform): fix stdlib unit test; guard vader llvm on Windows
- The emits_stdlib_calls test used `double` as a type name (Vader uses `float`);
  fixed. All 120 tests pass.
- `vader llvm` now fails fast on native Windows with a clear message (the native
  backend needs a POSIX runtime) pointing to WSL or the Go backend. Documented the
  platform-support matrix in SPEC.

+++ src/llvm.rs
@@ -2583,7 +2583,7 @@ mod tests {
                    []string p = strings.split(\"a,b\", \",\")\n\
                    string j = strings.join(p, \"-\")\n\
                    string u = strings.upper(p[0])\n\
-                   double r = math.sqrt(4.0)\n\
+                   float r = math.sqrt(4.0)\n\
                    }";
         let mut prog = parser::parse(lexer::tokenize(src).unwrap()).unwrap();
         let pkgs: HashSet<String> = ["strings".to_string(), "math".to_string()].into_iter().collect();
